requirements.txt
MouseInfo==0.1.3
PyGetWindow==0.0.9
PyRect==0.2.0
keyboard==0.13.5
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
pyautogui==0.9.54
pymsgbox==1.0.9
pyperclip==1.9.0
pyscreeze==1.0.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
